add: find-all?active

+++ src/main/java/com/construcpunto/managament_equipments/dto/viewLoanDto.java
@@ -1,11 +1,12 @@
 package com.construcpunto.managament_equipments.dto;
 
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class viewLoanDto {
 
-    private Long cedula;
+    private Integer cedula;
     private String clientName;
     private List<String> equipmentName;
     private Long promissoryNoteId;
@@ -14,11 +15,11 @@ public class viewLoanDto {
         this.equipmentName = new ArrayList<>();
     }
 
-    public Long getCedula() {
+    public Integer getCedula() {
         return cedula;
     }
 
-    public void setCedula(Long cedula) {
+    public void setCedula(Integer cedula) {
         this.cedula = cedula;
     }
